--- src/main/java/roomit/main/domain/config/test.java
@@ -1,4 +1,0 @@
-package roomit.main.domain.config;
-
-public class test {
-}
