+++ src/main/java/roomit/main/domain/member/entity/Member.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -14,8 +13,9 @@ import roomit.main.domain.member.entity.value.MemberEmail;
 import roomit.main.domain.member.entity.value.MemberNickname;
 import roomit.main.domain.member.entity.value.MemberPassword;
 import roomit.main.domain.member.entity.value.MemberPhoneNumber;
+import roomit.main.global.oauth2.dto.PROVIDER;
 import roomit.main.domain.reservation.entity.Reservation;
-import roomit.main.domain.review.entity.Review;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -58,6 +58,10 @@ public class Member {
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private LocalDate birthDay;
 
+    @Column(name = "provider")
+    @Enumerated(value = EnumType.STRING)
+    private PROVIDER provider;
+
     @CreatedDate
     private LocalDateTime createdAt;
 
@@ -67,7 +71,7 @@ public class Member {
     private List<Reservation> reservations = new ArrayList();
 
     @Builder
-    public Member(String memberNickName, String memberPhoneNumber, LocalDate birthDay, Sex memberSex, String memberEmail, String memberPwd, Role memberRole , PasswordEncoder passwordEncoder /*Reservation reservation*/) {
+    public Member(String memberNickName, String memberPhoneNumber, LocalDate birthDay, Sex memberSex, String memberEmail, String memberPwd, Role memberRole , PasswordEncoder passwordEncoder, PROVIDER provider /*Reservation reservation*/) {
         this.memberNickname = new MemberNickname(memberNickName);
         this.memberPhonenumber = new MemberPhoneNumber(memberPhoneNumber);
         this.memberEmail = new MemberEmail(memberEmail);
@@ -75,6 +79,7 @@ public class Member {
         this.birthDay = birthDay;
         this.memberSex = memberSex;
         this.memberRole = Role.ROLE_USER;
+        this.provider = provider;
 //        this.reservation = reservation;
     }
