+++ src/main/java/roomit/main/domain/member/repository/MemberRepository.java
@@ -3,6 +3,7 @@ package roomit.main.domain.member.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import roomit.main.domain.member.entity.Member;
+import roomit.main.global.oauth2.dto.PROVIDER;
 
 import java.util.Optional;
 
@@ -13,4 +14,7 @@ public interface MemberRepository extends JpaRepository<Member, Long> {
 
     @Query(value = "SELECT m FROM Member m WHERE m.memberNickname.value = :nickName")
     Member findByMemberNickName(String nickName);
+
+    @Query(value = "SELECT m FROM Member m WHERE m.memberNickname.value = :nickName AND m.provider = :provider")
+    Member findByMemberNickNameAndProvider(String nickName, PROVIDER provider);
 }
