+++ src/main/java/roomit/main/domain/payments/config/PaymentsConfig.java
@@ -20,5 +20,5 @@ public class PaymentsConfig {
     @Value("${payment.toss.fail_url}")
     private String failUrl;
 
-    public static final String URL = "https://api.tosspayments.com/v1/payments/";
+    public static final String URL = "https://api.tosspayments.com/v1/payments/confirm";
 }
