--- src/main/java/roomit/main/domain/token/controller/LoginController.java
@@ -1,111 +1,0 @@
-package roomit.main.domain.token.controller;
-
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-import roomit.main.domain.business.dto.CustomBusinessDetails;
-import roomit.main.domain.member.dto.CustomMemberDetails;
-import roomit.main.domain.token.config.JWTUtil;
-import roomit.main.domain.token.dto.LoginRequest;
-import roomit.main.domain.token.dto.LoginResponse;
-import roomit.main.domain.token.entity.RefreshEntity;
-import roomit.main.domain.token.repository.RefreshRepository;
-import roomit.main.global.config.security.util.CookieUtil;
-
-import java.util.Collections;
-import java.util.Date;
-
-@RestController
-@RequiredArgsConstructor
-public class LoginController {
-
-    private static final Logger log = LoggerFactory.getLogger(LoginController.class);
-    private final AuthenticationManager authenticationManager;
-    private final JWTUtil jwtUtil;
-    private final RefreshRepository refreshRepository;
-
-    @PostMapping("/login/member")
-    public ResponseEntity<?> memberLogin(@RequestBody LoginRequest loginRequest, HttpServletResponse response) {
-        try {
-            UsernamePasswordAuthenticationToken authenticationToken =
-                    new UsernamePasswordAuthenticationToken(loginRequest.getEmail(), loginRequest.getPassword(),Collections.emptyList());
-
-            Authentication authentication = authenticationManager.authenticate(authenticationToken);
-            CustomMemberDetails memberDetails = (CustomMemberDetails) authentication.getPrincipal();
-            String username = authentication.getName();
-
-            // Access Token 생성
-            String accessToken = jwtUtil.createJwt("access", memberDetails.getUsername(), "ROLE_USER", 1000 * 60 * 15L); // 15분 유효
-            // Refresh Token 생성
-            String refreshToken = jwtUtil.createJwt("refresh", memberDetails.getUsername(), "ROLE_USER", 1000 * 60 * 60 * 24L); // 24시간 유효
-
-            // Access Token은 헤더에 추가
-            response.addHeader("Authorization", "Bearer " + accessToken);
-
-            // Refresh Token을 쿠키에 저장
-            CookieUtil.addCookie(response, "refresh", refreshToken, 60 * 60 * 24); // 1일 유효
-
-            // Refresh Token을 DB에 저장
-            addRefreshEntity(username,refreshToken,6000000L);
-
-            // Access Token을 Body로 반환
-            return ResponseEntity.ok(new LoginResponse(accessToken));
-
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid credentials");
-        }
-    }
-
-    @PostMapping("/login/business")
-    public ResponseEntity<?> businessLogin(@RequestBody LoginRequest loginRequest, HttpServletResponse response) {
-        try {
-            UsernamePasswordAuthenticationToken authenticationToken =
-                    new UsernamePasswordAuthenticationToken(loginRequest.getEmail(), loginRequest.getPassword(), Collections.emptyList());
-
-            Authentication authentication = authenticationManager.authenticate(authenticationToken);
-            CustomBusinessDetails businessDetails = (CustomBusinessDetails) authentication.getPrincipal();
-            String username = authentication.getName();
-
-            // Access Token 생성
-            String accessToken = jwtUtil.createJwt("access", businessDetails.getUsername(), "ROLE_BUSINESS", 1000 * 60 * 15L); // 60 유효
-            // Refresh Token 생성
-            String refreshToken = jwtUtil.createJwt("refresh", businessDetails.getUsername(), "ROLE_BUSINESS", 1000 * 60 * 60 * 24L); // 24시간 유효
-
-            // Access Token은 헤더에 추가
-            response.addHeader("Authorization", "Bearer " + accessToken);
-
-            // Refresh Token을 쿠키에 저장
-            CookieUtil.addCookie(response, "refresh", refreshToken, 60 * 60 * 24); // 1일 유효
-
-            // Refresh Token을 DB에 저장
-            addRefreshEntity(username,refreshToken,6000000L);
-
-            // Access Token을 Body로 반환
-            return ResponseEntity.ok(new LoginResponse(accessToken));
-
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid credentials");
-        }
-    }
-
-    private void addRefreshEntity(String username, String refresh, Long expiredMs) {
-
-        Date date = new Date(System.currentTimeMillis() + expiredMs);
-
-        RefreshEntity refreshEntity = new RefreshEntity();
-        refreshEntity.setUsername(username);
-        refreshEntity.setRefresh(refresh);
-        refreshEntity.setExpiration(date.toString());
-
-        refreshRepository.save(refreshEntity);
-    }
-}
