--- src/main/java/roomit/main/domain/token/dto/LoginResponse.java
@@ -1,13 +1,0 @@
-package roomit.main.domain.token.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class LoginResponse {
-    private String token;
-
-}
