--- src/main/java/roomit/main/domain/token/entity/RefreshEntity.java
@@ -1,23 +1,0 @@
-package roomit.main.domain.token.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "refresh")
-public class RefreshEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String username;
-
-    private String refresh;
-
-    private String expiration;
-
-}
