--- src/main/java/roomit/main/domain/token/repository/RefreshRepository.java
@@ -1,13 +1,0 @@
-package roomit.main.domain.token.repository;
-
-import jakarta.transaction.Transactional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import roomit.main.domain.token.entity.RefreshEntity;
-
-public interface RefreshRepository extends JpaRepository<RefreshEntity,Long> {
-
-    Boolean existsByRefresh(String refresh);
-
-    @Transactional
-    void deleteByRefresh(String refresh);
-}
