+++ src/main/java/roomit/main/global/config/security/CorsMvcConfig.java
@@ -1,14 +1,18 @@
 package roomit.main.global.config.security;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 public class CorsMvcConfig implements WebMvcConfigurer {
 
+    @Value("${cors.url}")
+    private String corsUrl;
+
     @Override
     public void addCorsMappings(CorsRegistry corsRegistry) {
 
         corsRegistry.addMapping("/**")
-                .allowedOrigins("http://localhost:3000");
+                .allowedOrigins(corsUrl);
     }
 }
