+++ src/main/java/roomit/main/global/oauth2/dto/NaverResponse.java
@@ -1,0 +1,38 @@
+package roomit.main.global.oauth2.dto;
+
+import java.util.Map;
+
+public class NaverResponse implements OAuth2Response {
+
+    private final Map<String, Object> attribute;
+
+    public NaverResponse(Map<String, Object> attribute) {
+
+        this.attribute = (Map<String, Object>) attribute.get("response");
+    }
+
+    @Override
+    public PROVIDER getProvider() {
+
+        return PROVIDER.NAVER;
+    }
+
+    @Override
+    public String getProviderId() {
+
+        return attribute.get("id").toString();
+    }
+
+    @Override
+    public String getEmail() {
+
+        return attribute.get("email").toString();
+    }
+
+    @Override
+    public String getNickname() {
+
+        return attribute.get("nickname").toString();
+    }
+
+}
