+++ src/main/java/roomit/main/global/oauth2/dto/PROVIDER.java
@@ -1,0 +1,7 @@
+package roomit.main.global.oauth2.dto;
+
+public enum PROVIDER {
+    NAVER,
+    KAKAO,
+    BASIC
+}
