+++ src/main/java/roomit/main/global/token/dto/request/LoginRequest.java
@@ -1,0 +1,11 @@
+package roomit.main.global.token.dto.request;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class LoginRequest {
+    private String email;
+    private String password;
+}
