+++ src/main/java/roomit/main/global/token/dto/response/LoginResponse.java
@@ -1,0 +1,8 @@
+package roomit.main.global.token.dto.response;
+
+import roomit.main.domain.member.entity.Role;
+
+public record LoginResponse(
+        Role role
+) {
+}
