+++ src/main/java/roomit/main/global/token/dto/response/TokenResponse.java
@@ -1,0 +1,7 @@
+package roomit.main.global.token.dto.response;
+
+
+public record TokenResponse(
+        String token
+){}
+
